requirements.txt
# =========================================================
# CORE FRAMEWORK
# =========================================================
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1

# =========================================================
# FIREBASE (Auth, Firestore, Storage)
# =========================================================
firebase-admin==6.5.0
pyrebase5==5.0.1

# =========================================================
# UTILIDADES FLASK / WEB
# =========================================================
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==3.0.2
Werkzeug==3.0.1
requests==2.32.5
mercadopago==2.2.1
python-dateutil==2.9.0

# =========================================================
# MANEJO DE IMÁGENES
# =========================================================
Pillow==10.4.0

# =========================================================
# IA / MODELOS (CPU ONLY) - Windows compatible
# =========================================================
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# =========================================================
# UTILIDADES ADICIONALES
# =========================================================
python-multipart==0.0.6
email-validator==2.0.0
python-slugify==8.0.1

# =========================================================
# OPTIMIZACIÓN Y SEGURIDAD
# =========================================================
cachetools==5.3.1
certifi==2023.7.22
